requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langgraph==0.2.16
langchain-core>=0.2.27,<0.3.0
google-generativeai==0.3.2
pymongo==4.6.1
psutil==5.9.6
pytest==7.4.3
pytest-cov==4.1.0

